make toc section a button
