+++ app/styles/app.scss
@@ -66,7 +66,17 @@ table {
   padding-top: 1rem;
 }
 
-.level-1, .level-3 {
+.level-1 {
+  background-color: #ebeef2;
+  margin-top: 0.25rem;
+  padding: 0.65rem;
+  padding-left: 0;
+  color: #42474f;
+  font-size: 1.25rem;
+  text-align: left;
+}
+
+.level-3 {
   padding: 0.5em 0;
 }
